docs: record provider module refactor
