+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "eddie",
-    "version": "1.0.5",
+    "version": "1.0.6",
     "main": "dist/main.js",
     "bin": {
         "eddie": "dist/main.js"
