+++ src/core/engine/engine.module.ts
@@ -2,20 +2,27 @@ import { Module } from "@nestjs/common";
 import { ConfigModule } from "../../config/config.module";
 import { ContextModule } from "../context/context.module";
 import { IoModule } from "../../io";
-import { EngineService } from "./engine.service";
-import { ProviderFactory } from "../providers";
-import { ToolRegistryFactory } from "../tools/registry";
 import { HooksModule } from "../../hooks";
+import { ProvidersModule } from "../providers/providers.module";
 import { TokenizersModule } from "../tokenizers";
+import { EngineService } from "./engine.service";
+import { ToolRegistryFactory } from "../tools/registry";
 
 @Module({
-  imports: [ConfigModule, ContextModule, IoModule, HooksModule, TokenizersModule],
-  providers: [EngineService, ProviderFactory, ToolRegistryFactory],
+  imports: [
+    ConfigModule,
+    ContextModule,
+    IoModule,
+    HooksModule,
+    ProvidersModule,
+    TokenizersModule,
+  ],
+  providers: [EngineService, ToolRegistryFactory],
   exports: [
     EngineService,
-    ProviderFactory,
     ToolRegistryFactory,
     HooksModule,
+    ProvidersModule,
     TokenizersModule,
   ],
 })
